refactor(azure): keep parseAzureRepoUrl on AzureDevopsProvider
The extraction into a separate module was justified by an import cycle
that does not actually occur. Verified both ways:

- the compiled CLI resolves the static call fine, because it happens
  inside an async function long after every module is evaluated;
- importing the provider module directly still throws
  "Cannot access 'TicketProviderRoot' before initialization" - identically
  with and without this import. That breakage comes from the pre-existing
  ticketProviderRoot -> utils/index -> gitUtils -> ticketProvider/index
  cycle, which the extra module did nothing to avoid.

So the module bought nothing and cost a file. The parser goes back where
it has always lived, and the Azure Boards connector calls the static
method. Its tests move to azureDevopsRepoUrl.test.ts and now exercise it
through AzureDevopsProvider.
